move structs into deal
