+++ src/systems/filecoin_markets/deal/batch_process.go
@@ -1,0 +1,16 @@
+package deal
+
+type StorageDealPaymentAction = int
+
+const (
+	ExpireStorageDeals StorageDealPaymentAction = 0
+	CreditStorageDeals StorageDealPaymentAction = 1
+)
+
+type StorageDealSlashAction = int
+
+const (
+	SlashDeclaredFaults   StorageDealSlashAction = 0
+	SlashDetectedFaults   StorageDealSlashAction = 1
+	SlashTerminatedFaults StorageDealSlashAction = 2
+)
