+++ src/systems/filecoin_markets/deal/dealset.go
@@ -1,0 +1,31 @@
+package deal
+
+// TODO
+func (css *CompactDealSet) DealsOn() []DealID {
+	var dealId []DealID
+	return dealId
+}
+
+// TODO
+func (css *CompactDealSet) DealsOff() []DealID {
+	var dealId []DealID
+	return dealId
+}
+
+// TODO
+func (css *CompactDealSet) Add(dealId DealID) CompactDealSet {
+	var newCompactDealSet CompactDealSet
+	return newCompactDealSet
+}
+
+// TODO
+func (css *CompactDealSet) Remove(dealId DealID) CompactDealSet {
+	var newCompactDealSet CompactDealSet
+	return newCompactDealSet
+}
+
+// TODO
+func (css *CompactDealSet) Extend(css2 CompactDealSet) CompactDealSet {
+	var newCompactDealSet CompactDealSet
+	return newCompactDealSet
+}
